% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cortalconsors_transaction.R
\name{get_cortalconsors_sale}
\alias{get_cortalconsors_sale}
\title{Get sale transaction from Cortal Consors bank statement}
\usage{
get_cortalconsors_sale(df_pdf_page)
}
\arguments{
\item{df_pdf_page}{A data frame which results from the
function \code{\link[textreadr:read_pdf]{textreadr::read_pdf()}}.}
}
\value{
A data frame containing the parsed transaction.
}
\description{
Get sale transaction from Cortal Consors bank statement
}
