% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dkb_transaction.R
\name{get_dkb_transaction}
\alias{get_dkb_transaction}
\title{Get transaction from Deutsche Kreditbank (DKB AG) bank statement}
\usage{
get_dkb_transaction(df.pdf.page)
}
\arguments{
\item{df.pdf.page}{A data.frame which results from the function \code{\link[textreadr:read_pdf]{textreadr::read_pdf()}}.}
}
\description{
Get transaction from Deutsche Kreditbank (DKB AG) bank statement
}
