% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_onvista_transaction.R
\name{get_onvista_dividend}
\alias{get_onvista_dividend}
\title{Get dividend transaction from onVista bank statement}
\usage{
get_onvista_dividend(df.pdf.page)
}
\arguments{
\item{df.pdf.page}{A data frame which results from the function \code{\link[textreadr:read_pdf]{textreadr::read_pdf()}}.}
}
\description{
Get dividend transaction from onVista bank statement
}
