% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_onvista_transaction.R
\name{get_onvista_purchase}
\alias{get_onvista_purchase}
\title{Get purchase transaction from onVista bank statement}
\usage{
get_onvista_purchase(df.pdf.page)
}
\arguments{
\item{df.pdf.page}{A data.frame which results from the function \code{\link[textreadr:read_pdf]{textreadr::read_pdf()}}.}
}
\description{
Get purchase transaction from onVista bank statement
}
