% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_onvista_transaction.R
\name{get_onvista_vorabpauschale}
\alias{get_onvista_vorabpauschale}
\title{Get tax prepayment transaction from onVista bank statement}
\usage{
get_onvista_vorabpauschale(df.pdf.page)
}
\arguments{
\item{df.pdf.page}{A data.frame which results from the function \code{\link[textreadr:read_pdf]{textreadr::read_pdf()}}.}
}
\description{
Get tax prepayment transaction from onVista bank statement
}
