% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scalablecapital_transaction.R
\name{get_scalablecapital_purchase}
\alias{get_scalablecapital_purchase}
\title{Get purchase transaction from Scalable Capital statement}
\usage{
get_scalablecapital_purchase(df.pdf.page)
}
\arguments{
\item{df.pdf.page}{A data.frame which results from the function \code{\link[textreadr:read_pdf]{textreadr::read_pdf()}}.}
}
\value{
df.transaction.output A data.frame containing the parse transaction.
}
\description{
Get purchase transaction from Scalable Capital statement
}
