% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scalablecapital_transaction.R
\name{get_scalablecapital_transaction}
\alias{get_scalablecapital_transaction}
\title{Get transaction from Scalable Capital statement}
\usage{
get_scalablecapital_transaction(df.pdf.page)
}
\arguments{
\item{df.pdf.page}{A data frame which results from the function \code{\link[textreadr:read_pdf]{textreadr::read_pdf()}}.}
}
\value{
A data frame containing the parse transaction.
}
\description{
Get transaction from Scalable Capital statement
}
