% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_transaction_history.R
\name{init_transaction_history}
\alias{init_transaction_history}
\title{Create empty csv file for transaction history}
\usage{
init_transaction_history(path, file, overwrite = FALSE)
}
\arguments{
\item{path}{A single character string. The path of the csv file.}

\item{file}{A single character string. The name of the csv file.}

\item{overwrite}{Logical (default is FALSE) indicates whether csv shall be overwritten.}
}
\description{
Create empty csv file for transaction history
}
