% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_transaction_history.R
\name{update_transaction_history}
\alias{update_transaction_history}
\title{Update transaction history stored in a csv file}
\usage{
update_transaction_history(df_transactions, path, file)
}
\arguments{
\item{df_transactions}{A data frame which contains transactions
(e.g., results from \code{\link[=get_transactions]{get_transactions()}})}

\item{path}{A single character string. The path of the csv file.}

\item{file}{A single character string. The name of the csv file.}
}
\description{
Update transaction history stored in a csv file
}
