% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cortalconsors_transaction.R
\name{get_cortalconsors_transaction}
\alias{get_cortalconsors_transaction}
\title{Get transaction from Cortal Consors bank statement}
\usage{
get_cortalconsors_transaction(df.pdf.page)
}
\arguments{
\item{df.pdf.page}{A \code{data.frame} which results from the function \code{\link[textreadr:read_pdf]{textreadr::read_pdf()}}.}
}
\description{
Get transaction from Cortal Consors bank statement
}
