% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dkb_transaction.R
\name{get_dkb_transaction}
\alias{get_dkb_transaction}
\title{Get transaction from Deutsche Kreditbank (DKB AG) bank statement}
\usage{
get_dkb_transaction(df_pdf_page)
}
\arguments{
\item{df_pdf_page}{A data frame which results from the
function \code{\link[textreadr:read_pdf]{textreadr::read_pdf()}}.}
}
\value{
A data frame containing the parsed transaction.
}
\description{
Get transaction from Deutsche Kreditbank (DKB AG) bank statement
}
