% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_onvista_transaction.R
\name{get_onvista_purchase}
\alias{get_onvista_purchase}
\title{Get purchase transaction from onVista bank statement}
\usage{
get_onvista_purchase(df_pdf_page)
}
\arguments{
\item{df_pdf_page}{A data frame which results from the
function \code{\link[textreadr:read_pdf]{textreadr::read_pdf()}}.}
}
\value{
A data frame containing the parsed transaction.
}
\description{
Get purchase transaction from onVista bank statement
}
